Move api into internal/codeintel.
