+++ internal/codeintel/api/helpers_test.go
