+++ internal/codeintel/api/hover.go
